requirements.txt
google-cloud-documentai>=3.9.0
python-dotenv>=1.0.0